Flask jwt extended

+++ rest-api/resources/item.py
@@ -1,6 +1,13 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import get_jwt_identity, jwt_required, get_jwt_claims, fresh_jwt_required, jwt_optional
 from models.item import ItemModel
+
+
+"""
+The following resources contain endpoints that are protected by jwt,
+one may need a valid access token, a valid fresh token or a valid token with authorized privilege 
+to access each endpoint, details can be found in the README.md doc.  
+"""
 
 
 class Item(Resource):
@@ -16,29 +23,35 @@ class Item(Resource):
                         help="Every item needs a store_id."
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
 
-        data = Item.parser.parse_args()
+        data = self.parser.parse_args()
 
         item = ItemModel(name, **data)
 
         try:
             item.save_to_db()
         except:
-            return {"message": "An error occurred inserting the item."}, 500
+            return {"message": "An error occurred while inserting the item."}, 500
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -46,7 +59,7 @@ class Item(Resource):
         return {'message': 'Item not found.'}, 404
 
     def put(self, name):
-        data = Item.parser.parse_args()
+        data = self.parser.parse_args()
 
         item = ItemModel.find_by_name(name)
 
@@ -61,5 +74,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }, 200
